// src/api/types.gen.ts
export type BenchmarkMode = 'rating' | 'static' | 'live';
// ...
export type LlmByDifficulty = {
    name: string;
    provider: string;
    status: string;
    model_license?: string | null;
    organization?: string | null;
    rating?: number | string;
    display_name?: string | null;
    ratingEvent?: Array<RatingEvent>;
    validrate: number;
    passrate: number;
};
// ...
export type ProblemDifficulty = 'easy' | 'medium' | 'hard' | 'all';

export type RatingEvent = {
    rating: number;
    time: number;
    contestId?: number | null;
    name?: string | null;
    calcRating?: number | null;
};